refactor: split content index persistence
